Refactor workflow test support
